Remove http fluff from coap code. initialize coap context in ctor.
There was a lot of questionably designed parsing code for the HTTP transport. I decided not to really use its structure, particularly because we're using libcoap, and presumably it handles all of that.

The coap context should be initialized properly (assuming toCoapAddress( TTransport ) eventually works.

I presume that the coap_context is able to add resources dynamically. Hence the addition of addResource() to TCoapServer. Doing it this way just made life easier in terms of constructors, factories, etc.