Introduce conservative default timeout for TUdpSocket::avail()
Naturally, SOCK_DGRAM sockets are connectionless. Without getting into any technical details, that means that when the other side of the conversation decides to end the conversation, SOCK_DGRAM sockets receive no kind of notification.

That, and the fact that the default SO_RCVTIMEO in Thrift is 0 (infinite), meant that my unit tests would hang when there was no more data from the client side.

The TUdpSocket::avail() method is important to receive precisely the correct size of datagram from a UDP socket without the unnecessary baggage of large, static buffers.

With this change, avail() will not hang indefinitely if the other end of the conversation has "hung up".